+++ app/src/main/java/com/nailorsh/repeton/features/settings/UserSettingsRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package com.nailorsh.repeton.core.settings
+package com.nailorsh.repeton.features.settings
 
 import android.content.Context
 import androidx.datastore.core.DataStore
